Added profile button functionality#49
