+++ lib/Services/auth_wrapper.dart
@@ -1,5 +1,7 @@
 import 'package:batua/UI/home_screen.dart';
+import 'package:batua/UI/home_view.dart';
 import 'package:batua/UI/login_screen.dart';
+import 'package:batua/UI/profile.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -11,9 +13,9 @@ class AuthWrapper extends StatelessWidget {
 
     // return either the Home or Login screen
     if (user == null) {
-      return LoginScreen();
-    } else {
       return HomeScreen();
+    } else {
+      return Myprofile();
     }
   }
 }
